requirements.txt
# AI & Agents
anthropic>=0.40.0
mcp>=1.0.0

# Web Search
tavily-python>=0.3.0

# API & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=12.0
python-multipart>=0.0.12

# HTTP Client (agent-to-agent calls)
httpx>=0.27.0

# Data & Config
pydantic>=2.0.0
python-dotenv>=1.0.0

# File handling
aiofiles>=24.0.0
